complete thank you component
