--- src/components/FormStep5.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const FormStep5 = () => {
-  return (
-    <div>FormStep5</div>
-  )
-}
-
-export default FormStep5
